Refactor checkboxes and store
